pyproject.toml
[tool.poetry]
name = "kaziapi"
version = "0.1.0"
description = ""
authors = ["monzero <briankeny15@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
django-rest-framework = "^0.1.0"
rest-framework-simplejwt = "^0.0.2"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"
whitenoise = "^6.7.0"
dj-database-url = "^2.2.0"
pillow = "^10.4.0"
djangorestframework-simplejwt = "^5.3.1"
django-phonenumber-field = "^8.0.0"
phonenumbers = "^8.13.47"
africastalking = "^1.2.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
